lion_app/requirements.txt
Django>=4.2.3
gunicorn>=21.2.0
psycopg2-binary==2.9.7
pymongo==4.4.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
boto3==1.28.22
Pillow==10.0.0
